fix: ensure mismatch in component canvas size will not cause an err in diff algo
Since baselineShotImage and featureShotImage are expected to be the same
size, we need to ensure that they have the same dimensions, otherwise
pixelmatch will not generate the diff image correctly and will throw an
error: Image sizes do not match. we can fill in the missing pixels with
transparent pixels.